--- explore/src/test/java/com/amoalla/redditube/explore/ExploreApplicationTests.java
@@ -1,13 +1,0 @@
-package com.amoalla.redditube.explore;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ExploreApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
