+++ explorer/.gitignore
@@ -1,6 +1,6 @@
 HELP.md
 target/
-!../.mvn/wrapper/maven-wrapper.jar
+!.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**/target/
 !**/src/test/**/target/
