+++ explorer/pom.xml
@@ -8,7 +8,7 @@
         <version>0.0.1-SNAPSHOT</version>
     </parent>
 
-    <artifactId>explore</artifactId>
+    <artifactId>explorer</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>${project.artifactId}</name>
     <description>Service for exploring subreddits' and users' media posts</description>
