+++ explorer/src/main/resources/application.yml
@@ -2,7 +2,7 @@ spring:
   profiles:
     active: oauth
   application:
-    name: explore
+    name: explorer
 
 server:
   port: 8081
