+++ explorer/src/test/java/com/amoalla/redditube/explorer/ExplorerApplicationTests.java
@@ -1,0 +1,16 @@
+package com.amoalla.redditube.explorer;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+@SpringBootTest
+class ExplorerApplicationTests {
+
+	@Test
+	void contextLoads() {
+		assertTrue(true);
+	}
+
+}
